# openapi/kitchen-api.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "description": "Машиночитаемое описание ошибки.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "Единая оболочка ошибочного HTTP-ответа.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "MenuItemResponse": {
        "description": "Позиция меню; цена указана в минимальных единицах валюты.",
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_available": {
            "title": "Is Available",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "price_minor": {
            "exclusiveMinimum": 0.0,
            "title": "Price Minor",
            "type": "integer"
          },
          "restaurant_id": {
            "format": "uuid",
            "title": "Restaurant Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "restaurant_id",
          "name",
          "description",
          "price_minor",
          "is_available"
        ],
        "title": "MenuItemResponse",
        "type": "object"
      },
      "MenuResponse": {
        "description": "Страница меню заведения.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MenuItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationResponse"
          }
        },
        "required": [
          "items",
          "pagination"
        ],
        "title": "MenuResponse",
        "type": "object"
      },
...
      "PaginationResponse": {
        "description": "Метаданные offset-пагинации.",
        "properties": {
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "limit",
          "offset",
          "total"
        ],
        "title": "PaginationResponse",
        "type": "object"
      },
...
      "RestaurantListResponse": {
        "description": "Страница доступных заведений.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RestaurantResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationResponse"
          }
        },
        "required": [
          "items",
          "pagination"
        ],
        "title": "RestaurantListResponse",
        "type": "object"
      },
      "RestaurantResponse": {
        "description": "Заведение в пользовательском каталоге.",
        "properties": {
          "currency": {
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "currency"
        ],
        "title": "RestaurantResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "API платформы для заказа еды",
    "title": "Авито.Кухня API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/restaurants": {
      "get": {
        "operationId": "list_restaurants_api_v1_restaurants_get",
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 10000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestaurantListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Некорректный запрос"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Требуется авторизация"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Объект не найден"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Конфликт состояния"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Ошибка валидации запроса"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Внутренняя ошибка сервиса"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Сервис временно недоступен"
          }
        },
        "summary": "Получить список заведений",
        "tags": [
          "Каталог"
        ]
      }
    },
    "/api/v1/restaurants/{restaurant_id}/menu": {
      "get": {
        "operationId": "get_menu_api_v1_restaurants__restaurant_id__menu_get",
        "parameters": [
          {
            "in": "path",
            "name": "restaurant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Restaurant Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "available_only",
            "required": false,
            "schema": {
              "default": true,
              "title": "Available Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 10000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Некорректный запрос"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Требуется авторизация"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Заведение не найдено"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Конфликт состояния"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Ошибка валидации запроса"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Внутренняя ошибка сервиса"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Сервис временно недоступен"
          }
        },
        "summary": "Получить меню заведения",
        "tags": [
          "Каталог"
        ]
      }
    }
  }
}